lab2 Initial commit
